1.增加投票逻辑 2.修复banner中英文问题

+++ app/src/main/java/com/breadwallet/wallet/wallets/ela/request/Outputs.java
@@ -28,6 +28,6 @@ public class Outputs {
 
     public String addr;
 
-    public double amt;
+    public long amt;
 
 }
